requirements.txt
paramiko
psycopg[binary]
prometheus-client
requests
python-dotenv

pytest
ruff
bandit
pip-audit
